resolution of the issue Fichier Inversion.COMM  #5

+++ application/mini-LOMOS/format_ginette.sh
@@ -7,31 +7,36 @@
 # $4 : pressure sensor name
 # $5 : beginning of the experiment
 # $6 : time step (s)
-# $7 : thickness of the studied hyporheic zone (= distance between the 2 furthest working PT100, in meters)
-# $8 : nb of meshes between the river bed and the second PT100
-# $9 : nb of meshes between the river bed and the third PT100
-# $10 : nb of meshes between the river bed and the fourth PT100
-# $11 : nb of functioning PT100 (3 if every sensor works, 2 if one is broken, etc.)
-# $12 : Number of areas (clay, sand...)
-# $13 : Thickness of each area (separator = "", in meters)
+# $7 : nb of functioning PT100 (3 if every sensor works, 2 if one is broken, etc.)
+# $8 : Number of areas (clay, sand...)
+# $9 : Thickness of each area (separator = "", in meters)
 
 echo "-- Entering format_ginette.sh"
 ## Read calibrated parameter
 input_parameter="inversion_parameter.COMM"
 #fichier cree qu'on remplit
+
+
 tested_ranges="E_tested_ranges.dat"
 nbligne=$(wc -l $input_parameter |cut -d" " -f1)
-a=$((${12}*4));
-rm awk.ginette
+a=$((${8}*4));
+
+if [[ -f "$tested_ranges" ]]; then
+    rm awk.ginette
 rm E_tested_ranges.dat
-for i in `seq 1 $a`; do echo 1 1 1 >> awk.ginette; done
+fi
+if [[ -f E_nb_zone.dat ]]; then
 rm E_nb_zone.dat
-echo ${12}  ${13} >> E_nb_zone.dat
+fi
+
+for i in `seq 1 $a`; do echo 1 1 1 >> awk.ginette; done
+
+echo ${8}  ${9} >> E_nb_zone.dat
 
 if  [ $nbligne != $a ]
 then
         echo "Le nombre de zone et de parametres des fichiers de inversion_parameter COMM ne correspondent pas !"
-        echo  " ${12} zones  sont imposees" "inversion_parameter COMM contient" $nbligne "au lieu de" $a
+        echo  " ${8} zones  sont imposees" "inversion_parameter COMM contient" $nbligne "au lieu de" $a
 exit
 else
         echo "Le nombre de zone et de parametres des fichiers de inversion_parameter COMM correspondent !" $a
@@ -41,7 +46,7 @@ fi
 cp awk.ginette $tested_ranges
 while read  param zone min max pas
 do
-nbzone=${12}
+nbzone=${8}
 k="k"
 n="n"
 l="l"
@@ -71,47 +76,35 @@ done <$input_parameter
 
 chmod 755 *.sh
 #supprime tous les fichiers contenant "test_" dans le repertoire ./ (non recursif), ils feraient planter "formatToGinette.R"
-find -maxdepth 1 -name '*test_*.COMM' | xargs rm
-mkdir ./GINETTE_SENSI/OBS/
-rm -rf ./GINETTE_SENSI/OBS/*
-rm -rf ./GINETTE_SENSI/OUTPUT/*
+#find -maxdepth 1 -name '*test_*.COMM' | xargs rm
+OBS=./GINETTE_SENSI/OBS/
+if [ -d $OBS ]; then
+rm -r $OBS
+fi
+mkdir $OBS
+OUTPUT=./GINETTE_SENSI/OUTPUT
+if [ -d $OUTPUT ]; then
+rm -r $OUTPUT
+fi
+mkdir $OUTPUT
+
 cp $tested_ranges ./GINETTE_SENSI/
 
-## creation du fichier [test_nomdupoint.COMM]
-./format_ts.sh $1 $2 $3 $4 $5
-#time observation and simulation -> Comparision to adapt the boundary condition files
-echo $6 > tmp.txt
-cat tmp.txt test_$1.COMM > tmp2.txt
-mv tmp2.txt test_$1.COMM
-rm tmp.txt
 
 
 
-#Add the number of cell for the interpolation of the initial value in the file main.c
-L=$7
-nm=$(awk -v L=$L 'BEGIN { print ((L ) / 0.01) }');
-echo $nm > tmp.txt
-cat tmp.txt test_$1.COMM > tmp2.txt
-mv tmp2.txt test_$1.COMM
-rm tmp.txt
-
-
-echo $5 > tmp.txt
-cat tmp.txt test_$1.COMM > tmp2.txt
-mv tmp2.txt test_$1.COMM
-rm tmp.txt
 
 ## fichier [test_nomdupoint.COMM] cree
 ## Creation des fichiers d'entree pour Ginette
 R CMD BATCH formatToGinette.R
 
 ##fichier observation pour inversion
-for i in `seq 1 ${11}`;
+for i in `seq 1 ${10}`;
 do
 mv Obs_temperature_maille$i\_t.dat ./GINETTE_SENSI/OBS/Obs_temperature_maille$i\_t.dat
 done #done i
 
-for i in E_pression_initiale.dat E_charge_t.dat E_temp_t.dat E_temperature_initiale.dat nitt.dat
+for i in E_pression_initiale.dat E_charge_t.dat E_temp_t.dat E_temperature_initiale.dat nitt.dat model.dat
 do
 mv $i ./GINETTE_SENSI/
 done
@@ -120,12 +113,8 @@ mv E_nb_zone.dat ./GINETTE_SENSI/
 
 cd ./GINETTE_SENSI
 echo "Now preparing ginette in" `pwd`
-dos2unix *.dat
-#echo "compiling CmpKro"
-#cd ./CmpKro
-#make clean
-#make
-#cd ../
+dos2unix --quiet *.dat
+
 
 
 nitt=`awk '{print $NF}' nitt.dat`
@@ -134,23 +123,34 @@ mv awk.out E_parametre.dat
 
 
 
-pwd
-echo "Lenght of 1D model L=" $8
+#Add the number of cell and size of the colunm
+nm=`awk '{print $1}' model.dat`
+L=`awk '{print $2}' model.dat`
+nmaille1=`awk '{print $3}' model.dat`
+
+if (( ${10} > 1 )) ; then
+nmaille2=`awk '{print $4}' model.dat`
+else
+nmaille2=10
+fi
+if (( "${10}" > 2 )) ; then
+nmaille3=`awk '{print $5}' model.dat`
+else
+nmaille3=10
+fi
+
 awk -f CmdConfig_nm.awk -v nm="$nm" -v  L="$L" E_parametre.dat
 mv awk.out E_parametre.dat
-awk -f CmdConfig_nmaille.awk -v nmaille1="${8}" -v  nmaille2="${9}"  -v  nmaille3="${10}" E_parametre.dat
+awk -f CmdConfig_nmaille.awk -v nmaille1="$nmaille1" -v  nmaille2="$nmaille2"  -v  nmaille3="$nmaille3" E_parametre.dat
 mv awk.out E_parametre.dat
 
-#echo "Launching the sensitivity for" $1
-#./HZ1D.sh ${11} ${12}
+echo "Launching the sensitivity for" $1
+./HZ1D.sh ${10} ${8}
 cd ..
 # Calcul des critères stats
-if [ -d SENSI ]; then
-rm -r SENSI
-fi
+#if [ -d SENSI ]; then
+#rm -r SENSI
+#fi
 #mkdir SENSI
 #R CMD BATCH Comparaison_mailles_sim-obs.R
 
-# rm -rf Sim_*.dat
-#cp ./Sensi_final.dat ../$1_Sensi_final.dat
-#echo "Sensibility file moved in "$1"_Sensi_final.dat"
